pyproject.toml
[project]
name = "familiar-ai"
version = "0.1.0"
description = "AI that lives alongside you — with eyes, voice, and memory"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "onvif-zeep-async>=4.0.0",
    "Pillow>=10.0.0",
    "tinytuya>=1.15.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "openai-whisper",
    "sentence-transformers>=3.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
familiar = "familiar_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/familiar_agent"]

[tool.ruff]
line-length = 100
